requirements.txt
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.4
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2
django-allauth==0.58.2
django-celery-results==2.5.1
django-debug-toolbar==4.2.0
django-modeltranslation==0.18.11
django-rosetta==0.9.9
environs==9.5.0
gunicorn==20.0.4
html5lib==1.1
idna==3.4
kombu==5.3.2
lxml==4.9.3
marshmallow==3.20.1
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.2
Pillow==10.0.1
polib==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.3
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.8.0
pypdf==3.16.2
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.2
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.9
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.11

